refactor: split editor and canvas into smaller components
