+++ lib/uuid.ts
@@ -1,0 +1,5 @@
+export function uuid() {
+    return typeof crypto !== 'undefined' && 'randomUUID' in crypto
+        ? crypto.randomUUID()
+        : Math.random().toString(36).slice(2);
+}
